Rename UpdateTemplateScopeMutation to UpdateReflectTemplateScopeMutation
